src/main/java/uk/gov/hmcts/reform/bailcaseapi/domain/handlers/postsubmit/DocumentsUploadedConfirmation.java
package uk.gov.hmcts.reform.bailcaseapi.domain.handlers.postsubmit;

import org.springframework.stereotype.Component;
import uk.gov.hmcts.reform.bailcaseapi.domain.entities.BailCase;
import uk.gov.hmcts.reform.bailcaseapi.domain.entities.ccd.Event;
import uk.gov.hmcts.reform.bailcaseapi.domain.entities.ccd.callback.Callback;
import uk.gov.hmcts.reform.bailcaseapi.domain.entities.ccd.callback.PostSubmitCallbackResponse;
import uk.gov.hmcts.reform.bailcaseapi.domain.handlers.PostSubmitCallbackHandler;

@Component
public class DocumentsUploadedConfirmation implements PostSubmitCallbackHandler<BailCase> {

    @Override
    public boolean canHandle(Callback<BailCase> callback) {
        return (callback.getEvent() == Event.UPLOAD_DOCUMENTS);
    }

    @Override
    public PostSubmitCallbackResponse handle(Callback<BailCase> callback) {
        if (!canHandle(callback)) {
            throw new IllegalStateException("Cannot handle callback");
        }

        PostSubmitCallbackResponse postSubmitResponse =
            new PostSubmitCallbackResponse();
        postSubmitResponse.setConfirmationHeader("# Your upload is complete");
        postSubmitResponse.setConfirmationBody(
            "#### What happens next\n\n"
            + "The document(s) you uploaded are available to view in the [documents tab](/cases/case-details/"
            + callback.getCaseDetails().getId()
            + "#Documents)."
        );
        return postSubmitResponse;
    }
}
